--- a/src/Compilers/Lua/Portable/Parser/Blender.Reader.cs
+++ b/src/Compilers/Lua/Portable/Parser/Blender.Reader.cs
@@ -151,9 +151,9 @@ namespace Loretta.CodeAnalysis.Lua.Syntax.InternalSyntax
 
             private SyntaxToken LexNewToken()
             {
-                if (_lexer.Position != _newPosition)
+                if (_lexer.TextWindow.Position != _newPosition)
                 {
-                    _lexer.Restore(_newPosition);
+                    _lexer.TextWindow.Reset(_newPosition);
                 }
 
                 var token = _lexer.Lex();
@@ -218,7 +218,7 @@ namespace Loretta.CodeAnalysis.Lua.Syntax.InternalSyntax
 
                 // don't reuse nodes or tokens with skipped text or diagnostics attached to them
                 if (nodeOrToken.ContainsDiagnostics ||
-                    (nodeOrToken.IsToken && ((LuaSyntaxNode) nodeOrToken.AsToken().Node).ContainsSkippedText && nodeOrToken.Parent.ContainsDiagnostics))
+                    (nodeOrToken.IsToken && ((SyntaxTrivia) nodeOrToken.AsToken().Node).ContainsSkippedText && nodeOrToken.Parent.ContainsDiagnostics))
                 {
                     return false;
                 }