… .github/workflows/long-lexer-tests.yml
… src/Compilers/Lua/Portable/Parser/Lexer.QuickScanner.cs
… src/Compilers/Lua/Portable/Parser/Lexer.ShortString.cs
… src/Compilers/Lua/Portable/Parser/Lexer.cs
… src/Compilers/Lua/Portable/Parser/LexerCache.cs
… src/Compilers/Lua/Portable/Parser/SlidingTextWindow.cs
… src/Compilers/Lua/Portable/Parser/SourceTextReader.cs
… src/Compilers/Lua/Portable/Parser/SyntaxParser.cs
… src/Compilers/Lua/Portable/Syntax/InternalSyntax/SyntaxFactory.cs
… src/Compilers/Lua/Portable/Utilities/CharUtils.cs
… src/Compilers/Lua/Test/Portable/Lexical/LexicalTests.cs
… src/InternalBenchmarks/results/LexTimeBenchmark.md
… src/InternalBenchmarks/results/ParseTimeBenchmark.md
… src/Tools/Loretta.Generators.SyntaxXml/SourceWriter.cs