Optimizes the lexing process by bringing in a few more things from Roslyn such as the [`SlidingTextWindow`](https://github.com/dotnet/roslyn/blob/07c890cb76d8e62fa29553c84a23896feec3b78b/src/Compilers/CSharp/Portable/Parser/SlidingTextWindow.cs) and the [`QuickScanner.cs`](https://github.com/dotnet/roslyn/blob/07c890cb76d8e62fa29553c84a23896feec3b78b/src/Compilers/CSharp/Portable/Parser/QuickScanner.cs) state machine.

## Lexing Improvements
| File         | Previous Mean | Previous Throughput | Current Mean | Current Throughput |  Time Reduction | Throughput Increase |
|:-------------|--------------:|--------------------:|-------------:|-------------------:|----------------:|--------------------:|
|   `anim.lua` |      121.2 μs |         22.87 MiB/s |     30.80 μs |        89.95 MiB/s |      **74.59%** |            **393%** |
| `rustic.lua` |  435,721.3 μs |         13.78 MiB/s | 57,572.13 μs |       104.30 MiB/s |      **86.79%** |            **757%** |

## Parsing Improvements
| File         | Previous Mean | Previous Throughput |  Current Mean | Current Throughput |  Time Reduction | Throughput Increase |
|:-------------|--------------:|--------------------:|--------------:|-------------------:|----------------:|--------------------:|
|   `anim.lua` |      233.1 μs |         11.89 MiB/s |      58.31 μs |        47.51 MiB/s |      **74.98%** |            **399%** |
| `rustic.lua` |  713,273.8 μs |          8.42 MiB/s | 152,176.99 μs |        39.46 MiB/s |      **78.66%** |            **469%** |